config.mak.dev: add a DEVOPTS=analyzer mode to use GCC's -fanalyzer
Add an option to compile with GCC's -fanalyzer, which as noted in [1]
has become much more useful in the upcoming GCC 12 version.

Start by quieting various existing issues, many of which are flagging
genuine issues, but for now we can ensure that no more of these creep
in without our notice.

1. https://developers.redhat.com/articles/2022/04/12/state-static-analysis-gcc-12-compiler
2. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=105264

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>